android_webview/tools/system_webview_shell/test/data/blink-apis/webmidi/requestmidiaccess-with-sysex.html

<!DOCTYPE html>
<html>
  <script src="../../resources/js-test.js"></script>
  <body>
    <script type="text/javascript">
      description("Test navigator.requestMIDIAccess API in WebView");
      window.jsTestIsAsync = true;

      function onMIDIError(errorMessage) {
        testFailed('requestMIDIAccess error callback, reason: ' + errorMessage);
        finishJSTest();
      }

      function onMIDISuccess(access) {
        window.access = access;
        shouldBeTrue('access.sysexEnabled');
        testPassed('requestMIDIAccess with option sysex:true succeeded.');
        finishJSTest();
      }

      navigator.requestMIDIAccess({sysex:true}).then(
          onMIDISuccess, onMIDIError);
    </script>
  </body>
</html>
